+++ CHANGELOG.md
@@ -1,3 +1,1 @@
-- Added title in `DiagnosisScreen.kt`
-- `SymptomsActivity.kt` uploads diagnosis to Firestore
-- `UserPrescriptions.kt` checks for empty prescriptions now
+- Improved ProfileScreen in 'Profile.kt'